Make the dot pattern in the group project respond to the music
I chose to work on the audio part of the project. I used preload() in p5.js to load the music, and applied the FFT() function to make the rotation speed of the dot pattern respond to the music. I also added a button to control music playback and pause.